backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
